Fixed expansion issue on small screen sizes, mobile display now auto-renders
