+++ src/Components/FinderDisplay.jsx
@@ -58,6 +58,7 @@ import MetroLogo from '../images/MetroNewsLogo.png';
 import CANewsLogo from '../images/CANewsOttawaLogo.png';
 import OttawaTechLogo from '../images/OttawaTechWatchLogo.png';
 
+var ChartAnimate = require("./ChartAnimate.jsx");
 
 //Image icons to be displayed in left-most "groupsBar" (All images are displayed at once)
 var groupIconImages = [
@@ -107,9 +108,8 @@ class FinderDisplay extends Component {
     };
     this.changeState = this.changeState.bind(this);
     this.resetAnimation = this.resetAnimation.bind(this);
-
-
   }
+
 
   groupClick(i) {
     //This is activated when a group (Profile, Work, Projects etc ) is clicked
@@ -208,128 +208,23 @@ class FinderDisplay extends Component {
       })
 
     }  else if (i == 2) {
-      // console.log("expand window clicked");
-      var html = document.getElementsByTagName('html')[0];
+      console.log(this.state.expandFull);
 
       if (this.state.expandFull == 0) {
-        //SPAGHETTI CODE!!!
         this.setState({
           expandFull: 1,
         })
-        //Transition animation when the expand navigation button is clicked (green circle in top left of finder display)
-        document.getElementById("backShadow").style.webkitAnimationName = "totalTransform";
-        document.getElementById("backShadow").style.webkitAnimationDuration = "0.15s";
-        document.getElementsByClassName("bothWindows")[0].style.webkitAnimationName = "totalTransform";
-        document.getElementsByClassName("bothWindows")[0].style.webkitAnimationDuration = "0.15s";
-
-        document.getElementsByClassName("finderTopBar")[0].style.webkitAnimationName = "topTransform";
-        document.getElementsByClassName("finderTopBar")[0].style.webkitAnimationDuration = "0.15s";
-
-        document.getElementsByClassName("bottomElemeents")[0].style.webkitAnimationName = "bottomTransform";
-        document.getElementsByClassName("bottomElemeents")[0].style.webkitAnimationDuration = "0.15s";
-
-        document.getElementsByClassName("finderSideBar")[0].style.webkitAnimationName = "groupTransform";
-        document.getElementsByClassName("finderSideBar")[0].style.webkitAnimationDuration = "0.15s";
-
-        document.getElementsByClassName("finderFilesBar")[0].style.webkitAnimationName = "fileTransform";
-        document.getElementsByClassName("finderFilesBar")[0].style.webkitAnimationDuration = "0.15s";
-
-        document.getElementsByClassName("finderDescriptionBar")[0].style.webkitAnimationName = "descTransform";
-        document.getElementsByClassName("finderDescriptionBar")[0].style.webkitAnimationDuration = "0.15s";
-
-        //The code below transforms the chart. However, it crashes if there is a page without a chart so the conditional statement avoids this crash.
-        if (myInfo.info[this.state.groupIndex][this.state.fileIndex].length > 7) {
-          document.getElementsByTagName("dl")[0].style.webkitAnimationName = "chartTransform";
-          document.getElementsByTagName("dl")[0].style.webkitAnimationDuration = "0.15s";
-        }
-
-
-        for (var i = 0; i < myInfo.info[this.state.groupIndex].length;i++) {
-          document.getElementsByClassName("fileStyle")[i].style.webkitAnimationName = "fileStyleTransform";
-          document.getElementsByClassName("fileStyle")[i].style.webkitAnimationDuration = "0.15s";
-          document.getElementsByClassName("fileStyle2")[i].style.webkitAnimationName = "fileStyleTransform";
-          document.getElementsByClassName("fileStyle2")[i].style.webkitAnimationDuration = "0.15s";
-        }
-        for (var i = 0; i < myInfo.info.length;i++) {
-          document.getElementsByClassName("groupStyle")[i].style.webkitAnimationName = "groupStyleTransform";
-          document.getElementsByClassName("groupStyle")[i].style.webkitAnimationDuration = "0.15s";
-          document.getElementsByClassName("groupStyle2")[i].style.webkitAnimationName = "groupStyleTransform";
-          document.getElementsByClassName("groupStyle2")[i].style.webkitAnimationDuration = "0.15s";
-        }
-        html.style.setProperty("--totalWidth", "75vw");
-        html.style.setProperty("--totalHeight", "90vh");
-        html.style.setProperty("--topWidth",  "75vw");
-        html.style.setProperty("--topHeight", "10vh");
-        html.style.setProperty("--groupWidth", "15vw");
-        html.style.setProperty("--groupHeight", "75vh");
-        html.style.setProperty("--fileWidth", "15vw");
-        html.style.setProperty("--fileHeight", "75vh");
-        html.style.setProperty("--descWidth", "45vw");
-        html.style.setProperty("--descHeight", "75vh");
-        html.style.setProperty("--chartWidth", "35vw");
-        setTimeout(this.resetAnimation, 150);
-
       } else {
-        //EVEN MORE SPAGHETTI CODE!!!!
-        html.style.setProperty("--totalWidth", "600px");
-        html.style.setProperty("--totalHeight", "450px");
-        html.style.setProperty("--topWidth", "600px");
-        html.style.setProperty("--topHeight", "70px");
-        html.style.setProperty("--groupWidth", "130px");
-        html.style.setProperty("--groupHeight", "380px");
-        html.style.setProperty("--fileWidth", "180px");
-        html.style.setProperty("--fileHeight", "380px");
-        html.style.setProperty("--descWidth", "290px");
-        html.style.setProperty("--descHeight", "380px");
-        html.style.setProperty("--chartWidth", "140px");
-        setTimeout(this.resetAnimation, 150);
-        // console.log("minimizng");
         this.setState({
           expandFull: 0,
         })
-        document.getElementById("backShadow").style.webkitAnimationName = "totalTransformRev";
-        document.getElementById("backShadow").style.webkitAnimationDuration = "0.15s";
-
-        document.getElementsByClassName("bothWindows")[0].style.webkitAnimationName = "totalTransformRev";
-        document.getElementsByClassName("bothWindows")[0].style.webkitAnimationDuration = "0.15s";
-
-        document.getElementsByClassName("finderTopBar")[0].style.webkitAnimationName = "topTransformRev";
-        document.getElementsByClassName("finderTopBar")[0].style.webkitAnimationDuration = "0.15s";
-
-        document.getElementsByClassName("bottomElemeents")[0].style.webkitAnimationName = "bottomTransformRev";
-        document.getElementsByClassName("bottomElemeents")[0].style.webkitAnimationDuration = "0.15s";
-
-        document.getElementsByClassName("finderSideBar")[0].style.webkitAnimationName = "groupTransformRev";
-        document.getElementsByClassName("finderSideBar")[0].style.webkitAnimationDuration = "0.15s";
-
-        document.getElementsByClassName("finderFilesBar")[0].style.webkitAnimationName = "fileTransformRev";
-        document.getElementsByClassName("finderFilesBar")[0].style.webkitAnimationDuration = "0.15s";
-
-        document.getElementsByClassName("finderDescriptionBar")[0].style.webkitAnimationName = "descTransformRev";
-        document.getElementsByClassName("finderDescriptionBar")[0].style.webkitAnimationDuration = "0.15s";
-
-        //The code below transforms the chart. However, it crashes if there is a page without a chart so the conditional statement avoids this crash.
-        if (myInfo.info[this.state.groupIndex][this.state.fileIndex].length > 7) {
-          document.getElementsByTagName("dl")[0].style.webkitAnimationName = "chartTransformRev";
-          document.getElementsByTagName("dl")[0].style.webkitAnimationDuration = "0.15s";
-        }
-        for (var i = 0; i < myInfo.info[this.state.groupIndex].length;i++) {
-          document.getElementsByClassName("fileStyle")[i].style.webkitAnimationName = "fileStyleTransformRev";
-          document.getElementsByClassName("fileStyle")[i].style.webkitAnimationDuration = "0.15s";
-          document.getElementsByClassName("fileStyle2")[i].style.webkitAnimationName = "fileStyleTransformRev";
-          document.getElementsByClassName("fileStyle2")[i].style.webkitAnimationDuration = "0.15s";
-        }
-
-        for (var i = 0; i < myInfo.info.length;i++) {
-          document.getElementsByClassName("groupStyle")[i].style.webkitAnimationName = "groupStyleTransformRev";
-          document.getElementsByClassName("groupStyle")[i].style.webkitAnimationDuration = "0.15s";
-          document.getElementsByClassName("groupStyle2")[i].style.webkitAnimationName = "groupStyleTransformRev";
-          document.getElementsByClassName("groupStyle2")[i].style.webkitAnimationDuration = "0.15s";
-        }
-
-
-
       }
+      if (this.state.width <= 800) {
+        console.log("mobile detected");
+
+        this.state.mobileDisplay = 1;
+      }
+      ChartAnimate(this.state);
     }
   }
